fix: this solution move package.json outside the app folder
