+++ common/classes/public-inputs.mjs
@@ -5,7 +5,7 @@ Class to construct a public input hash (PIH).  Unlike a commitment, the PIH
 will vary quite a bit and so we use an array as the input
 */
 import gen from 'general-number';
-import sha256 from '../utils/crypto/sha256.mjs';
+import sha256 from '../../app/src/utils/crypto/sha256.mjs';
 
 const { generalise } = gen;
