+++ common/classes/transaction.mjs
@@ -4,8 +4,8 @@
 An optimistic Transaction class
 */
 import gen from 'general-number';
-import Web3 from '../utils/web3.mjs';
-import { compressProof } from '../utils/curve-maths/curves.mjs';
+import Web3 from '../../app/src/utils/web3.mjs';
+import { compressProof } from '../../app/src/utils/curve-maths/curves.mjs';
 
 const { generalise } = gen;
