+++ nightfall-client/src/classes/index.mjs
@@ -1,9 +1,8 @@
 import Commitment from './commitment.mjs';
 import Nullifier from './nullifier.mjs';
 
-// below file are common file and volume shared via docker-compose.yml
-import Proof from './proof.mjs'; // eslint-disable-line import/no-unresolved
-import PublicInputs from './public-inputs.mjs'; // eslint-disable-line import/no-unresolved
-import Transaction from './transaction.mjs'; // eslint-disable-line import/no-unresolved
+import Proof from '../../../common/classes/proof.mjs';
+import PublicInputs from '../../../common/classes/public-inputs.mjs';
+import Transaction from '../../../common/classes/transaction.mjs';
 
 export { Proof, PublicInputs, Transaction, Commitment, Nullifier };
