+++ nightfall-optimist/.gitignore
@@ -102,8 +102,3 @@ dist
 
 # TernJS port file
 .tern-port
-
-# empty files created in src from docker-compose volume
-src/classes/proof.mjs
-src/classes/public-inputs.mjs
-src/classes/transaction.mjs
