+++ nightfall-optimist/Dockerfile
@@ -11,8 +11,11 @@ WORKDIR /app
 RUN mkdir /app/mongodb
 
 COPY src src
-COPY docker-entrypoint.sh package.json package-lock.json ./
+COPY docker-entrypoint.sh ./
 RUN chmod 755 docker-entrypoint.sh
+
+WORKDIR /
+COPY package.json package-lock.json ./
 
 RUN npm ci
