+++ nightfall-optimist/package.json
@@ -4,8 +4,8 @@
   "description": "A container for optimistic proposals and challenges",
   "main": "index.mjs",
   "scripts": {
-    "start": "node src/index.mjs",
-    "dev": "nodemon src/index.mjs"
+    "start": "node ${WORKDIR}src/index.mjs",
+    "dev": "nodemon --watch ${WORKDIR}src ${WORKDIR}src/index.mjs"
   },
   "repository": {
     "type": "git",
@@ -22,9 +22,7 @@
   },
   "homepage": "https://github.com/EYBlockchain/nightfall-optimist#readme",
   "nodemonConfig": {
-    "ignore": [
-      "build/*"
-    ]
+    "verbose": true
   },
   "dependencies": {
     "axios": "^0.21.1",
