+++ nightfall-optimist/src/classes/index.mjs
@@ -4,10 +4,9 @@ import Proposer from './proposer.mjs';
 import TransactionError from './transaction-error.mjs';
 import VerificationKey from './verification-key.mjs';
 
-// below file are common file and volume shared via docker-compose.yml
-import Proof from './proof.mjs'; // eslint-disable-line import/no-unresolved
-import PublicInputs from './public-inputs.mjs'; // eslint-disable-line import/no-unresolved
-import Transaction from './transaction.mjs'; // eslint-disable-line import/no-unresolved
+import Proof from '../../../common/classes/proof.mjs';
+import PublicInputs from '../../../common/classes/public-inputs.mjs';
+import Transaction from '../../../common/classes/transaction.mjs';
 
 export {
   Proof,
